animations & surrounded map with trees w/ collision. no battle stuff yet.

+++ src/common.cpp
@@ -1,18 +1,18 @@
-#include "util.hpp"
+#include "common.hpp"
 
-sf::Vector2f directionToUnitVector(Direction d) {
+sf::Vector2i directionToUnitVector(Direction d) {
 	switch (d) {
 	case Direction::Left:
-		return sf::Vector2f(-1.0, 0.0);
+		return sf::Vector2i(-1, 0);
 	case Direction::Right:
-		return sf::Vector2f(1.0, 0.0);
+		return sf::Vector2i(1, 0);
 	case Direction::Up:
-		return sf::Vector2f(0.0, -1.0);
+		return sf::Vector2i(0, -1);
 	case Direction::Down:
-		return sf::Vector2f(0.0, 1.0);
+		return sf::Vector2i(0, 1);
 	case Direction::None:;
 	}
-	return sf::Vector2f(0.0, 0.0);
+	return sf::Vector2i(0, 0);
 }
 
 int toMultiple(int round, int multiple) {
